Added tcp socket implamentation. Implamented peer connector, which establishes connections with given peers and stores them for use

+++ BitTorrentEdu/DTOs/Peer.cs
@@ -4,7 +4,7 @@ namespace BitTorrentEdu.DTOs
 {
     public class Peer
     {
-        public Peer(IPAddress ip, long port, string id = null)
+        public Peer(IPAddress ip, int port, string id = null)
         {
             Id = id;
             Ip = ip;
@@ -13,6 +13,6 @@ namespace BitTorrentEdu.DTOs
 
         public string Id { get; private set; }
         public IPAddress Ip { get; private set; }
-        public long Port { get; private set; }
+        public int Port { get; private set; }
     }
 }
